feat(readable page title): fix test title output

+++ packages/application-shell/src/components/application-page-title/application-page-title.spec.tsx
@@ -51,7 +51,7 @@ describe.each`
 describe.each`
   content                                                                                                      | title
   ${[`Short title`]}                                                                                           | ${`Short title - Products - my-project-key - Merchant Center`}
-  ${[`Long title, a very long product name`]}                                                                  | ${`Short title,...product name - Products - my-project-key - Merchant Center`}
+  ${[`Long title, a very long product name`]}                                                                  | ${`Long title, ...product name - Products - my-project-key - Merchant Center`}
   ${['Short title1', 'Short title2']}                                                                          | ${`Short title1 - Short title2 - Products - my-project-key - Merchant Center`}
   ${['First title of product, a very long product name', 'Second title of product, a very long product name']} | ${`First title ...product name - Second title...product name - Products - my-project-key - Merchant Center`}
 `(`'when the custom "content" is "$content"`, ({ content, title }) => {
